feat: 🎸 Improved variable definition analysis
Added context_id field to identifier. Make define function in context
return the context index where the variable was defined. Renamed
is_defined to functoin to get__context and modified it so that it
returns the context index. Moved tests to a separate folder and added
test for the context indices